requirements.txt
# Hands-free Voice Dictation - Requirements

# Local speech-to-text (GPU-accelerated)
faster-whisper>=1.0.0
torch>=2.0.0

# Audio capture
sounddevice>=0.5.0
numpy>=1.24.0

# High-quality resampling (optional, falls back to linear interpolation)
samplerate>=0.2.1

# System tray indicator (optional)
pystray>=0.19.5
Pillow>=10.0.0

# Voxtral API mode (optional, for cloud transcription)
# mistralai[realtime]>=1.6.0

# Push-to-talk / toggle hotkeys (only needed for dictate.py mode)
# pynput>=1.7.6
